# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dtos.CreateCategoryRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      parent_id:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
  dtos.CreateInventoryRequest:
    properties:
      max_quantity:
        type: integer
      min_quantity:
        minimum: 0
        type: integer
      product_id:
        type: string
      quantity:
        minimum: 0
        type: integer
      store_id:
        type: string
      warehouse_id:
        type: string
    required:
    - max_quantity
    - min_quantity
    - product_id
    - quantity
    - store_id
    - warehouse_id
    type: object
...
  dtos.UpdateCategoryRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      parent_id:
        type: string
      slug:
        type: string
    type: object
  dtos.UpdateInventoryRequest:
    properties:
      product_id:
        type: string
      quantity:
        minimum: 0
        type: integer
      status:
        enum:
        - in_stock
        - out_of_stock
        - low_stock
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: integer
      success:
        type: boolean
    type: object
host: localhost:2707
info:
  contact:
    email: support@luka-platform.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Luka Platform e-commerce application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Luka Platform API
  version: "1.0"
paths:
...
  /categories:
    post:
      consumes:
      - application/json
      description: Create a new category with the provided details
      parameters:
      - description: Category details
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category with the provided ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Fetch a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get a category by ID
      tags:
      - categories
    patch:
      consumes:
      - application/json
      description: Update a category with the provided details
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category details
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - categories
  /inventories:
    post:
      consumes:
      - application/json
      description: Create a new inventory with the provided details
      parameters:
      - description: Inventory details
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateInventoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create a new inventory
      tags:
      - inventories
  /inventories/{id}:
    delete:
      description: Delete an inventory from the system
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete an inventory
      tags:
      - inventories
    get:
      description: Get detailed information about a specific inventory
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get inventory by ID
      tags:
      - inventories
    patch:
      consumes:
      - application/json
      description: Update an existing inventory
      parameters:
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      - description: Inventory update data
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateInventoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update an inventory
      tags:
      - inventories
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"